chore: commit before pull
