+++ src/main/java/com/sejong/aistudyassistant/quiz/QuizController.java
@@ -21,6 +21,18 @@ public class QuizController {
         this.jwtUtil=jwtUtil;
     }
 
+    @PostMapping("/create")
+    public ResponseEntity<Void> createQuiz(@RequestHeader("Authorization") String authHeader,@RequestBody CreateQuizRequest quizRequest) {
+        String token = authHeader.replace("Bearer ", "");
+        Long userId=jwtUtil.getUserIdFromToken(token);
+        try {
+            quizService.createQuizzes(quizRequest.lectureText(), quizRequest.summaryId(), userId);
+            return ResponseEntity.status(200).body(null);
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(null);
+        }
+    }
+
     @GetMapping("/learning")
     public ResponseEntity<List<GetLearningQuizResponse>> getLearningQuiz(@RequestHeader("Authorization") String authHeader,@RequestBody GetLearningQuizRequest quizRequest) {
 
@@ -70,4 +82,5 @@ public class QuizController {
             return ResponseEntity.status(500).body(null);
         }
     }
+
 }
