--- src/main/java/com/sejong/aistudyassistant/quiz/dto/GetSubjectStatisticsResponse.java
@@ -1,9 +1,0 @@
-package com.sejong.aistudyassistant.quiz.dto;
-
-public record GetSubjectStatisticsResponse (
-
-        String subjectName,
-        int totalQuestions,
-        int correctAnswers
-
-){}
