+++ src/main/java/com/sejong/aistudyassistant/summary/dto/SummaryResponseDto.java
@@ -1,0 +1,26 @@
+package com.sejong.aistudyassistant.summary.dto;
+
+public class SummaryResponseDto {
+
+    private Long id;
+    private String summaryText;
+    private Long transcriptId;
+
+    public SummaryResponseDto(Long id, String summaryText, Long transcriptId) {
+        this.id=id;
+        this.summaryText = summaryText;
+        this.transcriptId = transcriptId;
+
+    }
+
+    public Long getId() {
+        return id;
+    }
+    public String getSummaryText() {
+        return summaryText;
+    }
+
+    public Long getTranscriptId() {
+        return transcriptId;
+    }
+}
