+++ src/main/resources/application.properties
@@ -10,8 +10,6 @@ spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
 openai.api.key=${OPENAI_API_KEY}
 daglo.api.token=${DAGLO_API_TOKEN}
-
-
-jwt.secret=your_new_long_secret_key_of_at_least_32_characters
+jwt.secret=${JWT_SECRET_KEY}
